feat(v0.3.1): Consolidate ListenerAgent into CustomAgent
Merges ListenerAgent functionality into CustomAgent for simpler API.

BREAKING CHANGES:
- ListenerAgent profile removed
- Replace `from jarviscore.profiles import ListenerAgent` with `CustomAgent`
- All handler methods (on_peer_request, on_peer_notify) work identically

What's New:
- CustomAgent now includes built-in P2P handlers
- Simplified profile architecture (AutoAgent + CustomAgent only)
- Handler-based pattern is now the recommended approach
- Backward compatibility: run() method still works

Updated:
- README.md: Removed ListenerAgent references, updated examples
- docs/: Updated all documentation (CUSTOMAGENT_GUIDE, GETTING_STARTED, etc.)
- examples/: Renamed listeneragent_* → customagent_*, updated all examples
- tests/: Updated test imports and class names
- jarviscore/__init__.py: Removed ListenerAgent export
- jarviscore/profiles/: Removed listeneragent.py, updated customagent.py
- pyproject.toml: Bumped version to 0.3.1

Added:
- customagent_cognitive_discovery_example.py: Replacement example
