+++ jarviscore/__init__.py
@@ -4,16 +4,15 @@ JarvisCore - P2P Distributed Agent Framework
 A production-grade framework for building autonomous agent systems with:
 - P2P coordination via SWIM protocol
 - Workflow orchestration with dependencies
-- Three agent profiles: AutoAgent, CustomAgent, and ListenerAgent
+- Two agent profiles: AutoAgent and CustomAgent
 
 Profiles:
-    AutoAgent     - LLM generates and executes code from prompts (autonomous mode)
-    CustomAgent   - You provide execute_task() or run() (p2p/distributed modes)
-    ListenerAgent - API-first agents with background P2P (just implement handlers)
+    AutoAgent   - LLM generates and executes code from prompts (autonomous mode)
+    CustomAgent - You provide handlers or execute_task() (p2p/distributed modes)
 
 Modes:
     autonomous  - Workflow engine only (AutoAgent)
-    p2p         - P2P coordinator only (CustomAgent/ListenerAgent with run() loops)
+    p2p         - P2P coordinator only (CustomAgent with run() loop)
     distributed - Both workflow + P2P (CustomAgent with execute_task())
 
 Quick Start (AutoAgent - autonomous mode):
@@ -30,12 +29,12 @@ Quick Start (AutoAgent - autonomous mode):
     await mesh.start()
     results = await mesh.workflow("calc", [{"agent": "calculator", "task": "Calculate 10!"}])
 
-Quick Start (ListenerAgent + FastAPI):
+Quick Start (CustomAgent + FastAPI):
     from fastapi import FastAPI
-    from jarviscore.profiles import ListenerAgent
+    from jarviscore.profiles import CustomAgent
     from jarviscore.integrations.fastapi import JarvisLifespan
 
-    class MyAgent(ListenerAgent):
+    class MyAgent(CustomAgent):
         role = "processor"
         capabilities = ["processing"]
 
@@ -61,7 +60,7 @@ Quick Start (CustomAgent - distributed mode):
     results = await mesh.workflow("demo", [{"agent": "processor", "task": "hello"}])
 """
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 __author__ = "JarvisCore Contributors"
 __license__ = "MIT"
 
@@ -73,7 +72,6 @@ from jarviscore.core.mesh import Mesh, MeshMode
 # Execution profiles
 from jarviscore.profiles.autoagent import AutoAgent
 from jarviscore.profiles.customagent import CustomAgent
-from jarviscore.profiles.listeneragent import ListenerAgent
 
 # Custom Profile: Decorator, Wrapper, and Context
 from jarviscore.adapter import jarvis_agent, wrap
@@ -99,7 +97,6 @@ __all__ = [
     # Profiles
     "AutoAgent",
     "CustomAgent",
-    "ListenerAgent",
 
     # Custom Profile (decorator and wrapper)
     "jarvis_agent",
