+++ jarviscore/profiles/__init__.py
@@ -3,12 +3,10 @@ Execution profiles for agents.
 
 Profiles define HOW agents execute tasks:
 - AutoAgent: LLM-powered code generation + sandboxed execution
-- CustomAgent: User-defined logic (LangChain, MCP, raw Python)
-- ListenerAgent: API-first agents with background P2P listening
+- CustomAgent: User-defined logic with P2P message handling
 """
 
 from .autoagent import AutoAgent
 from .customagent import CustomAgent
-from .listeneragent import ListenerAgent
 
-__all__ = ["AutoAgent", "CustomAgent", "ListenerAgent"]
+__all__ = ["AutoAgent", "CustomAgent"]
